requirements.txt
requests>=2.31.0
pandas>=2.1.0
openpyxl>=3.1.2
python-dotenv>=1.0.0
pymongo>=4.6.1
numpy>=1.24.0  # Required by pandas
typing-extensions>=4.7.0  # For enhanced type hints
python-dateutil>=2.8.2  # Required for datetime handling
